refactor(mm-next): adjust index page related components into `/index`
